+++ CHANGELOG.md
@@ -10,12 +10,16 @@ Format as below (Requires the '### Added' as a header or task fails)
 ...- Update ReadMe.MD
 
 ## [Unreleased]
+### Added
+- Build-ObjectType Add description parameter
+- New-InsightObjectTypeAttributes Add description parameter
 
 ## [1.14.20210413.0] - 2021-04-13
 ### Added
 - Fixed bug in zoom room builder where parent_room was not created or populated. 
 - Added field 'typeValue' for type 'Object' in New-InsightObjectTypeAttributes
 - Added Find-InsightObjects
+- Added Find-InsightObjectsAdvanced
 - Removed Links from JSON and Build-ObjectType
 
 ## [1.13.20210412.0] - 2021-04-12